test: 메뉴 그룹 서비스 테스트 분리
